+++ internal/dmsg-discovery/api/api.go
@@ -22,6 +22,7 @@ import (
 	"github.com/skycoin/dmsg/internal/discmetrics"
 	"github.com/skycoin/dmsg/internal/dmsg-discovery/store"
 	"github.com/skycoin/dmsg/pkg/disc"
+	"github.com/skycoin/dmsg/pkg/dmsg"
 )
 
 var log = logging.MustGetLogger("dmsg-discovery")
@@ -43,10 +44,13 @@ type API struct {
 	startedAt                   time.Time
 	enableLoadTesting           bool
 	dmsgAddr                    string
+	DmsgServers                 []string
+	authPassphrase              string
+	OfficialServers             map[string]bool
 }
 
 // New returns a new API object, which can be started as a server
-func New(log logrus.FieldLogger, db store.Storer, m discmetrics.Metrics, testMode, enableLoadTesting, enableMetrics bool, dmsgAddr string) *API {
+func New(log logrus.FieldLogger, db store.Storer, m discmetrics.Metrics, testMode, enableLoadTesting, enableMetrics bool, dmsgAddr, authPassphrase string) *API {
 	if log != nil {
 		log = logging.MustGetLogger("dmsg_disc")
 	}
@@ -65,6 +69,9 @@ func New(log logrus.FieldLogger, db store.Storer, m discmetrics.Metrics, testMod
 		enableLoadTesting:           enableLoadTesting,
 		reqsInFlightCountMiddleware: metricsutil.NewRequestsInFlightCountMiddleware(),
 		dmsgAddr:                    dmsgAddr,
+		DmsgServers:                 []string{},
+		authPassphrase:              authPassphrase,
+		OfficialServers:             make(map[string]bool),
 	}
 
 	r.Use(middleware.RequestID)
@@ -249,7 +256,6 @@ func (a *API) setEntry() func(w http.ResponseWriter, r *http.Request) {
 		if timeout := r.URL.Query().Get("timeout"); timeout == "true" {
 			entryTimeout = store.DefaultTimeout
 		}
-
 		entry := new(disc.Entry)
 		if err := json.NewDecoder(r.Body).Decode(entry); err != nil {
 			a.handleError(w, r, disc.ErrUnexpected)
@@ -282,6 +288,14 @@ func (a *API) setEntry() func(w http.ResponseWriter, r *http.Request) {
 			if err := entry.VerifySignature(); err != nil {
 				a.handleError(w, r, disc.ErrUnauthorized)
 				return
+			}
+		}
+
+		if entry.Server != nil {
+			if entry.Server.ServerType == a.authPassphrase || a.OfficialServers[entry.Static.Hex()] {
+				entry.Server.ServerType = dmsg.DefaultOfficialDmsgServerType
+			} else {
+				entry.Server.ServerType = dmsg.DefaultCommunityDmsgServerType
 			}
 		}
 
@@ -415,9 +429,10 @@ func (a *API) getAllServers() http.HandlerFunc {
 func (a *API) serviceHealth(w http.ResponseWriter, r *http.Request) {
 	info := buildinfo.Get()
 	a.writeJSON(w, r, http.StatusOK, httputil.HealthCheckResponse{
-		BuildInfo: info,
-		StartedAt: a.startedAt,
-		DmsgAddr:  a.dmsgAddr,
+		BuildInfo:   info,
+		StartedAt:   a.startedAt,
+		DmsgAddr:    a.dmsgAddr,
+		DmsgServers: a.DmsgServers,
 	})
 }